src/sonar3d/wlsonar/pyproject.toml
[project]
name = "wlsonar"
version = "0.2.0"
description = "Python client and Range Image Protocol utilities for Water Linked Sonar 3D-15."
readme = "README.md"
authors = [
    { name = "Water Linked" }
]
requires-python = ">=3.10"
dependencies = [
    "protobuf>=6.33.2",
    "python-snappy>=0.7.3",
    "requests>=2.32.5",
]

[project.urls]
Homepage = "https://www.waterlinked.com/3dsonar"
Repository = "https://github.com/waterlinked/wlsonar"

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[tool.mypy]
warn_return_any = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "mypy-protobuf>=4.0.0",
    "pdbpp>=0.11.7",
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[[tool.mypy.overrides]]
module = ["requests.*", "snappy.*"]
follow_untyped_imports = true

[tool.ruff]
line-length = 100
exclude = [
    "*_pb2.py",
    "*_pb2.pyi",
]


[tool.ruff.format]
quote-style = "double"
line-ending = "auto"


[tool.ruff.lint]
select = [
    "D",   # pydocstyle
    "F",   # pyflakes
    "E",   # pycodestyle
    "I",   # isort
    "ANN", # flake8 type annotations
    "RUF", # ruff-specific rules
]
fixable = ["ALL"]
ignore = [
    "D107", # undocumented-public-init
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D"]
"test_*.py" = ["D"]


[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests that require real sonar hardware. Skipped by default.",
]
addopts = "-m 'not e2e'"
